Rename visibility option
